src/tukifact-ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tukifact-ai"
version = "0.1.0"
description = "TukiFact AI Agents — Validación, Clasificación y Extracción"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "nats-py>=2.9",
    "httpx>=0.27",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]
